# Config: Switch from `jsonschema` to `pydantic`