--- a/aiida/manage/configuration/options.py
+++ b/aiida/manage/configuration/options.py
@@ -18,9 +18,10 @@ __all__ = ('get_option', 'get_option_names', 'parse_option', 'Option')
 class Option:
     """Represent a configuration option schema."""
 
-    def __init__(self, name: str, schema: Dict[str, Any]):
+    def __init__(self, name: str, schema: Dict[str, Any], field):
         self._name = name
         self._schema = schema
+        self._field = field
 
     def __str__(self) -> str:
         return f'Option(name={self._name})'
@@ -30,26 +31,27 @@ class Option:
         return self._name
 
     @property
+    def valid_type(self) -> Any:
+        return self._field.type_
+
+    @property
     def schema(self) -> Dict[str, Any]:
         return self._schema
 
     @property
-    def valid_type(self) -> Any:
-        return self._schema.get('type', None)
-
-    @property
     def default(self) -> Any:
-        return self._schema.get('default', None)
+        return self._field.default
 
     @property
     def description(self) -> str:
-        return self._schema.get('description', '')
+        return self._field.field_info.description
 
     @property
     def global_only(self) -> bool:
-        return self._schema.get('global_only', False)
+        from .config import ProfileOptionsSchema
+        return self._name in ProfileOptionsSchema.__fields__
 
-    def validate(self, value: Any, cast: bool = True) -> Any:
+    def validate(self, value: Any) -> Any:
         """Validate a value
 
         :param value: The input value
@@ -59,68 +61,28 @@ class Option:
         :raise: ConfigValidationError
 
         """
-        # pylint: disable=too-many-branches
-        import jsonschema
+        value, validation_error = self._field.validate(value, {}, loc=None)
 
-        from aiida.manage.caching import _validate_identifier_pattern
-
-        from .config import ConfigValidationError
-
-        if cast:
-            try:
-                if self.valid_type == 'boolean':
-                    if isinstance(value, str):
-                        if value.strip().lower() in ['0', 'false', 'f']:
-                            value = False
-                        elif value.strip().lower() in ['1', 'true', 't']:
-                            value = True
-                    else:
-                        value = bool(value)
-                elif self.valid_type == 'string':
-                    value = str(value)
-                elif self.valid_type == 'integer':
-                    value = int(value)
-                elif self.valid_type == 'number':
-                    value = float(value)
-                elif self.valid_type == 'array' and isinstance(value, str):
-                    value = value.split()
-            except ValueError:
-                pass
-
-        try:
-            jsonschema.validate(instance=value, schema=self.schema)
-        except jsonschema.ValidationError as exc:
-            raise ConfigValidationError(message=exc.message, keypath=[self.name, *(exc.path or [])], schema=exc.schema)
-
-        # special caching validation
-        if self.name in ('caching.enabled_for', 'caching.disabled_for'):
-            for i, identifier in enumerate(value):
-                try:
-                    _validate_identifier_pattern(identifier=identifier)
-                except ValueError as exc:
-                    raise ConfigValidationError(message=str(exc), keypath=[self.name, str(i)])
+        if validation_error:
+            raise ConfigurationError(validation_error)
 
         return value
 
 
-def get_schema_options() -> Dict[str, Dict[str, Any]]:
-    """Return schema for options."""
-    from .config import config_schema
-    schema = config_schema()
-    return schema['definitions']['options']['properties']
-
-
 def get_option_names() -> List[str]:
     """Return a list of available option names."""
-    return list(get_schema_options())
+    from .config import GlobalOptionsSchema
+    return [key.replace('__', '.') for key in GlobalOptionsSchema.__fields__]
 
 
 def get_option(name: str) -> Option:
     """Return option."""
-    options = get_schema_options()
-    if name not in options:
+    from .config import GlobalOptionsSchema
+    options = GlobalOptionsSchema.__fields__
+    option_name = name.replace('.', '__')
+    if option_name not in options:
         raise ConfigurationError(f'the option {name} does not exist')
-    return Option(name, options[name])
+    return Option(name, GlobalOptionsSchema.schema()['properties'][option_name], options[option_name])
 
 
 def parse_option(option_name: str, option_value: Any) -> Tuple[Option, Any]:
@@ -132,6 +94,6 @@ def parse_option(option_name: str, option_value: Any) -> Tuple[Option, Any]:
 
     """
     option = get_option(option_name)
-    value = option.validate(option_value, cast=True)
+    value = option.validate(option_value)
 
     return option, value