--- a/docs/source/nitpick-exceptions
+++ b/docs/source/nitpick-exceptions
@@ -149,6 +149,15 @@ py:class ndarray
 py:class paramiko.proxy.ProxyCommand
 
 py:class pydantic.main.BaseModel
+py:class ModelPrivateAttr
+py:class CoreSchema
+py:class _decorators.DecoratorInfos
+py:class _generics.PydanticGenericMetadata
+py:class SchemaSerializer
+py:class SchemaValidator
+py:class Signature
+py:class ConfigDict
+py:class FieldInfo
 
 # These can be removed once they are properly included in the `__all__` in `plumpy`
 py:class plumpy.ports.PortNamespace