--- a/environment.yml
+++ b/environment.yml
@@ -26,7 +26,7 @@ dependencies:
 - pgsu~=0.2.1
 - psutil~=5.6
 - psycopg2-binary~=2.8
-- pydantic~=1.10
+- pydantic~=2.4
 - pytz~=2021.1
 - pyyaml~=6.0
 - requests~=2.0