@sphuber 2023-09-07
Note that this currently is working against `pydantic==1.10`. Ideally I want to use v2, which has been out for a while, but some indirect dependencies upstream have not migrated yet and so we cannot be compatible with it just yet. But they should be coming close. Probably we will update this PR to use v2 before merging it. This would be nice because pydantic v2 should be a lot faster, which matters since the parsing of the config will happen in almost all code pathways of AiiDA, so performance is important.