@sphuber 2023-09-07
@danielhollas this would get rid of `jsonschema`. However, it does introduce `pydantic`. Note that we want to add this for other reasons than `jsonschema` putting a load on import and runtime, but I think it would still be good to have an idea of the impact of `pydantic`. Would you be able to run your benchmark script on this branch and compare to current `main`? I plan to also add a version of this branch that uses pydantic v2 which should be a lot faster in runtime, but I don't know about import time.