@danielhollas 2023-09-07
@sphuber sure, I'll take a look, thanks! I was curious and already looked at import time of `pydantic` v2, and while it was about 30ms faster than `jsonschema` it was unfortunately still not negligible. I'll take a closer look.