@danielhollas 2023-09-12
Hi @sphuber,

Would you mind merging current main into this branch so that I can do the timings with the just merged import improvements?

Note that I am on a workshop this week, so unless there is a rush I'll do it next week.