@sphuber 2023-09-12
> Would you mind merging current main into this branch so that I can do the timings with the just merged import improvements?

Done. Thanks a lot. The first commit is against pydantic v1, and the second commit updates to v2. Might be interesting to see the difference between the two. Even though we almost certainly will have to go for v2.

> Note that I am on a workshop this week, so unless there is a rush I'll do it next week.

No worries