@danielhollas 2023-09-18
Unfortunately, pydantic v2 seems to cause a sizeable import time regression. :-( 

## main branch

```python
Benchmark 1: python -c 'import aiida'
  Time (mean ± σ):      90.4 ms ±   4.5 ms    [User: 80.1 ms, System: 10.2 ms]
  Range (min … max):    83.2 ms … 100.3 ms    31 runs

Benchmark 1: python -c 'import aiida.orm'
  Time (mean ± σ):     285.7 ms ±   7.0 ms    [User: 824.0 ms, System: 712.9 ms]
  Range (min … max):   278.6 ms … 299.7 ms    10 runs

Benchmark 1: python -c 'import aiida.cmdline.commands'
  Time (mean ± σ):     212.1 ms ±  12.5 ms    [User: 191.5 ms, System: 20.1 ms]
  Range (min … max):   199.4 ms … 238.2 ms    12 runs
```

## pydantic v2

```python
(aiida-dev) hollas-atmospec:~/atmospec/aiida-core (feature/pydantic)
17:31 $ hyperfine -w 3 "python -c 'import aiida'"
Benchmark 1: python -c 'import aiida'
  Time (mean ± σ):     183.4 ms ±   6.9 ms    [User: 166.4 ms, System: 16.8 ms]
  Range (min … max):   174.9 ms … 198.2 ms    16 runs
 
Benchmark 1: python -c 'import aiida.orm'
  Time (mean ± σ):     392.0 ms ±   9.1 ms    [User: 907.7 ms, System: 733.4 ms]
  Range (min … max):   378.3 ms … 405.7 ms    10 runs
 
Benchmark 1: python -c 'import aiida.cmdline.commands'
  Time (mean ± σ):     274.3 ms ±  10.7 ms    [User: 253.7 ms, System: 20.0 ms]
  Range (min … max):   263.9 ms … 298.0 ms    10 runs
```

## pydantic v1

```python
Benchmark 1: python -c 'import aiida'
  Time (mean ± σ):     135.3 ms ±   7.8 ms    [User: 121.8 ms, System: 13.3 ms]
  Range (min … max):   112.2 ms … 146.2 ms    21 runs

Benchmark 1: python -c 'import aiida.orm'
  Time (mean ± σ):     330.1 ms ±   6.7 ms    [User: 857.7 ms, System: 722.0 ms]
  Range (min … max):   320.9 ms … 345.7 ms    10 runs
 
Benchmark 1: python -c 'import aiida.cmdline.commands'
  Time (mean ± σ):     219.6 ms ±  14.3 ms    [User: 199.2 ms, System: 19.8 ms]
  Range (min … max):   196.8 ms … 240.9 ms    13 runs
```

## fastjsonschema branch

```python
Benchmark 1: python -c 'import aiida'
  Time (mean ± σ):      91.4 ms ±   4.3 ms    [User: 79.8 ms, System: 11.5 ms]
  Range (min … max):    83.6 ms …  98.8 ms    32 runs
 
(aiida-dev) hollas-atmospec:~/atmospec/aiida-core (fastjsonschema)
17:39 $ hyperfine -w 3 "python -c 'import aiida.orm'"
Benchmark 1: python -c 'import aiida.orm'
  Time (mean ± σ):     290.9 ms ±  11.4 ms    [User: 814.9 ms, System: 727.7 ms]
  Range (min … max):   276.1 ms … 305.5 ms    10 runs
 
(aiida-dev) hollas-atmospec:~/atmospec/aiida-core (fastjsonschema)
17:39 $ hyperfine -w 3 "python -c 'import aiida.cmdline.commands'"
Benchmark 1: python -c 'import aiida.cmdline.commands'
  Time (mean ± σ):     176.9 ms ±  12.3 ms    [User: 160.2 ms, System: 16.2 ms]
  Range (min … max):   160.2 ms … 196.5 ms    15 runs
```