@ltalirz 2023-09-22
possibly relevant thread https://github.com/pydantic/pydantic/discussions/6748