@sphuber 2023-09-22
> seems to me it makes sense to move this migration to pydantic v2 on hold until they figure out performance

We are anyway still blocked by https://github.com/materialsproject/emmet/pull/790
That being said, is a 50 ms difference really a blocker? Especially since it is a one-off when you first import the module. To me doesn't seem like a blocking issue.