@ltalirz 2023-09-22
I see a 90ms difference for importing `aiida`, doubling import time.

These timings are not important for running AiiDA scripts, it's always about the responsiveness of the verdi cli