@sphuber 2023-10-23
I am planning to merge this soon. @edan-bainglass did you still want to give this a look?

@danielhollas I have updated the branch onto the lastest `main` such that it contains your latest changes to the CLI speed. I get the following timings now for loading `verdi` (made sure to uninstall `trogon`):

### `main`
```
(aiida-py311) sph@invader:~/code/aiida/env/dev/aiida-core$ hyperfine -w 3 verdi
Benchmark 1: verdi
  Time (mean ± σ):     107.3 ms ±  26.2 ms    [User: 86.9 ms, System: 12.0 ms]
  Range (min … max):    84.4 ms … 170.4 ms    32 runs
```

### This PR
```
(aiida-py311) sph@invader:~/code/aiida/env/dev/aiida-core$ hyperfine -w 3 verdi
Benchmark 1: verdi
  Time (mean ± σ):     162.0 ms ±  18.4 ms    [User: 133.4 ms, System: 15.9 ms]
  Range (min … max):   135.8 ms … 204.8 ms    16 runs
```

There is a ~50 ms slowdown due to the changes. Note though that this PR also contains an update to `sqlalchemy` v2.0, which may be responsible for some, but I think most of it is probably `pydantic`. I am sorry to be undoing some of the hard-work you have been doing on the import times, but I think this slow-down is acceptable for the benefits that it will give us. To me the tab-completion feels snappy enough in any case.