@sphuber 2023-10-24
## `main`
(aiida-py311) sph@invader:~/code/aiida/env/dev/aiida-core$ hyperfine -w 3 "verdi profile list"
Benchmark 1: verdi profile list
  Time (mean ± σ):     149.4 ms ±  19.7 ms    [User: 123.4 ms, System: 14.2 ms]
  Range (min … max):   129.8 ms … 195.7 ms    18 runs

## PR
(aiida-py311) sph@invader:~/code/aiida/env/dev/aiida-core$ hyperfine -w 3 "verdi profile list"
Benchmark 1: verdi profile list
  Time (mean ± σ):     221.3 ms ±  28.2 ms    [User: 191.8 ms, System: 12.3 ms]
  Range (min … max):   197.6 ms … 281.0 ms    15 runs

Slightly bigger difference ~70 ms. I see quite a bit of variation in timings though, despite the warmup. So not sure how reliable these benchmarks are.